+++ pipis/__init__.py
@@ -69,9 +69,20 @@ def list_installed():
 
 
 @cli.command()
+@click.option('-r', '--requirement')
+@click.argument('name', nargs=-1, type=click.STRING)
 @click.confirmation_option()
-@click.argument('name', nargs=-1, required=True, type=click.STRING)
-def install(name):
+def install(requirement, name):
+    if not (requirement or name):
+        raise click.UsageError('missing arguments/options')
+    if requirement and name:
+        raise click.UsageError('too much arguments/options')
+    if requirement:
+        try:
+            with open(requirement, 'r') as req:
+                name = map(str.strip, req.readlines())
+        except IOError:
+            raise click.FileError(requirement)
     with click.progressbar(name, label='Installing', item_show_func=_show_package) as packages:
         for package in packages:
             venv_dir = os.path.join(VENV_ROOT_DIR, package)